+++ CHANGES.md
@@ -5,6 +5,7 @@
 - BREAKING CHANGE: 'create_decoder' has been renamed to 'decoder'
 - BREAKING CHANGE: 'create' has been renamed to 'router'
 - BREAKING CHANGE: 't' type has been renamed to `router`
+- BREAKING CHANGE: Remove labelled args for `decoder`
 - FIX: Fix matching query components in uri
 - FIX: Fix pretty printing of query components in uri
 - NEW: Add path, query and uri combinators to create and manipulate uri values